@@ -1,16 +1,16 @@
 module github.com/rbee3u/dpass
 
-go 1.22.0
+go 1.24.1
 
 require (
-	github.com/spf13/cobra v1.8.0
-	golang.org/x/crypto v0.19.0
-	golang.org/x/term v0.17.0
+	github.com/spf13/cobra v1.9.1
+	golang.org/x/crypto v0.36.0
+	golang.org/x/term v0.30.0
 	rsc.io/qr v0.2.0
 )
 
 require (
 	github.com/inconshreveable/mousetrap v1.1.0 // indirect
-	github.com/spf13/pflag v1.0.5 // indirect
-	golang.org/x/sys v0.17.0 // indirect
+	github.com/spf13/pflag v1.0.6 // indirect
+	golang.org/x/sys v0.31.0 // indirect
 )
