@@ -2,17 +2,13 @@ package solana
 
 import (
 	"crypto/ed25519"
-	"crypto/hmac"
-	"crypto/sha512"
-	"encoding/binary"
 	"errors"
 	"fmt"
 	"os"
-	"unsafe"
 
 	"github.com/rbee3u/dpass/internal/dcoin"
-	"github.com/rbee3u/dpass/third_party/github.com/mr-tron/base58"
-	"github.com/rbee3u/dpass/third_party/github.com/tyler-smith/go-bip32"
+	"github.com/rbee3u/dpass/pkg/base58"
+	"github.com/rbee3u/dpass/pkg/bip3x"
 	"github.com/spf13/cobra"
 )
 
@@ -22,8 +18,6 @@ const (
 	accountDefault = 0
 	changeDefault  = 0
 	secretDefault  = false
-
-	secretSize = 32
 )
 
 var (
@@ -31,8 +25,6 @@ var (
 	errInvalidCoin    = errors.New("invalid coin")
 	errInvalidAccount = errors.New("invalid account")
 	errInvalidChange  = errors.New("invalid change")
-
-	errNonHardenedChild = errors.New("non hardened child")
 )
 
 type backend struct {
@@ -56,32 +48,26 @@ func backendDefault() *backend {
 func NewCmd() *cobra.Command {
 	backend := backendDefault()
 	cmd := &cobra.Command{Use: "solana", Args: cobra.NoArgs, RunE: backend.runE}
-
 	cmd.Flags().Uint32Var(&backend.account, "account", accountDefault, fmt.Sprintf(
 		"account is the number of address (default %v)", accountDefault))
 	cmd.Flags().BoolVar(&backend.secret, "secret", secretDefault, fmt.Sprintf(
 		"show secret instead of address (default %t)", secretDefault))
-
 	return cmd
 }
 
 func (b *backend) checkArguments() error {
-	if b.purpose >= bip32.FirstHardenedChild {
+	if b.purpose >= bip3x.FirstHardenedChild {
 		return errInvalidPurpose
 	}
-
-	if b.coin >= bip32.FirstHardenedChild {
+	if b.coin >= bip3x.FirstHardenedChild {
 		return errInvalidCoin
 	}
-
-	if b.account >= bip32.FirstHardenedChild {
+	if b.account >= bip3x.FirstHardenedChild {
 		return errInvalidAccount
 	}
-
-	if b.change >= bip32.FirstHardenedChild {
+	if b.change >= bip3x.FirstHardenedChild {
 		return errInvalidChange
 	}
-
 	return nil
 }
 
@@ -90,16 +76,13 @@ func (b *backend) runE(_ *cobra.Command, _ []string) error {
 	if err != nil {
 		return fmt.Errorf("failed to read mnemonic: %w", err)
 	}
-
 	result, err := b.getResult(mnemonic)
 	if err != nil {
 		return fmt.Errorf("failed to get result: %w", err)
 	}
-
 	if _, err := os.Stdout.WriteString(result); err != nil {
 		return fmt.Errorf("failed to write result: %w", err)
 	}
-
 	return nil
 }
 
@@ -107,55 +90,22 @@ func (b *backend) getResult(mnemonic string) (string, error) {
 	if err := b.checkArguments(); err != nil {
 		return "", fmt.Errorf("failed to check arguments: %w", err)
 	}
-
-	seed, err := dcoin.MnemonicToSeed(mnemonic, "")
+	seed, err := bip3x.MnemonicToSeed(mnemonic, "")
 	if err != nil {
 		return "", fmt.Errorf("failed to convert mnemonic to seed: %w", err)
 	}
-
-	key, err := seedToKey(seed, []uint32{
-		bip32.FirstHardenedChild + b.purpose,
-		bip32.FirstHardenedChild + b.coin,
-		bip32.FirstHardenedChild + b.account,
-		bip32.FirstHardenedChild + b.change,
+	sk, err := bip3x.Ed25519DeriveSk(seed, []uint32{
+		b.purpose + bip3x.FirstHardenedChild,
+		b.coin + bip3x.FirstHardenedChild,
+		b.account + bip3x.FirstHardenedChild,
+		b.change + bip3x.FirstHardenedChild,
 	})
 	if err != nil {
-		return "", fmt.Errorf("failed to derive key from mnemonic: %w", err)
+		return "", fmt.Errorf("failed to derive sk: %w", err)
 	}
-
-	privateKey := ed25519.NewKeyFromSeed(key)
-
+	privateKey := ed25519.NewKeyFromSeed(sk)
 	if b.secret {
 		return base58.Encode(privateKey), nil
 	}
-
 	return base58.Encode(privateKey[ed25519.SeedSize:]), nil
-}
-
-func seedToKey(seed []byte, path []uint32) ([]byte, error) {
-	hash := hmac.New(sha512.New, []byte("ed25519 seed"))
-	hash.Write(seed)
-	digest := hash.Sum(nil)
-
-	for index := range path {
-		if path[index] < bip32.FirstHardenedChild {
-			return nil, errNonHardenedChild
-		}
-
-		hash = hmac.New(sha512.New, digest[secretSize:])
-		hash.Write([]byte{0})
-		hash.Write(digest[:secretSize])
-		hash.Write(uint32ToBytes(path[index]))
-		digest = hash.Sum(nil)
-	}
-
-	return digest[:secretSize], nil
-}
-
-func uint32ToBytes(v uint32) []byte {
-	b := make([]byte, unsafe.Sizeof(v))
-
-	binary.BigEndian.PutUint32(b, v)
-
-	return b
 }
